move scannerv4 client to pkg
