+++ pkg/scanners/scannerv4/client/mocks/client.go
